added lvsize to filesystem | create lv
This change makes the use of host_fs_layout more useful.
Different filesystems for audit and adr is a very good idea and possible with this change.
Defined a default of 14GB for /u01 in orahost/defaults/main.yml.
Please change your existing host_fs_layout defininition and add the lvsize entry.
The a1 and a2 defines the order of creating and mounting the filesystems.

Examples with ADR:
   host_fs_layout:
       a1:
         {mntp: /u01, device: /dev/sdb, vgname: vgora, pvname: /dev/sdb1, lvname: lvora, lvsize: 10G, fstype: ext4 }
       a2:
         {mntp: /u01/app/oracle/diag, device: /dev/sdb, vgname: vgora, pvname: /dev/sdb1, lvname: lvdiag, lvsize: 2G, fstype: ext4 }