preparation for using different ORACLE_HOME for each Database
This change is the 1st preparation which is needed for the possibility
to define a custom ORACLE_HOME for each database.

The following examples shows the usage of a custom ORACLE_HOME for the
racdb database. The value from home is ignored but the variable is requirred!
oracle_version_db is requirred for choosing the correct installation
media.

Example:
  oracle_databases:
        - home: dbhome_1
          oracle_db_name: racdb
          oracle_home: /u01/app/oracle/product/12.1.0.2/dbhome_1
          oracle_version_db: 12.1.0.2