hostroles for compatibility against SLES12.1
ansible-oracle gets compatibility against Suse Linux Enterprise Server.
Currently only 12.1 is supported.

The Firewall could not be disabled at the moment. You have to disable it
before using ansible-oracle. Normaly the installation of SLES12.1 disables
the sshd from outside and a change in the firewall is requirred. Please
disable it at this point.

Known Issues:
- missing oracleasm-support on SLES
SuSe includes the kernel modules for ASMlib but the oracleasm-support is
missing on the installation media. Please install it before using ansible-oracle
with enabled ASM support.
The RPM can be downloaded from Oracle:
http://www.oracle.com/technetwork/server-storage/linux/asmlib/sles12-2714894.html