This addresses #38 This change will make it possible to break out the passwords to an external file that can be easily encrypted, and is broken up into a GI part & a DB part.
GI:
- default_gipass: This will be the default password if either of the 2 following are not defined
- sysasmpassword: Password for the sysasm user
- asmmonitorpassword:  Password for the asmsnmp account

DB:
- default_dbpass: This will be the default/fallback password if DB specific passwords are not defined

The following is the password lookup-structure that can be defined per database:
dbpasswords:
       db1:                    <- This defines the passwords for the db1 database. If omitted all passwords falls back to default_dbpass
         sys: Oracle_456
         system: Oracle_456
         dbsnmp: Oracle_456
         pdbadmin: Oracle_456
       db2:                    <- This defines the passwords for the db2 database. If omitted all passwords falls back to default_dbpass
         sys: Oracle_789
         system: Oracle_789
         dbsnmp: Oracle_789
         pdbadmin: Oracle_789

As an interim solution to provide some backwards compatibility, the default passwords: default_dbpass & default_gipass, will look for the
passwords in the old structure and use that (if they are defined).

default_gipass: "{% if oracle_password is defined %}{{ oracle_password }}{% else %}Oracle123{% endif%}"
sysasmpassword: "{% if oracle_password is defined %}{{ oracle_password }}{% else %}Oracle123{% endif%}"
asmmonitorpassword: "{% if oracle_password is defined %}{{ oracle_password }}{% else %}Oracle123{% endif%}"
default_dbpass: "{% if item.0.oracle_db_passwd is defined %}{{ item.0.oracle_db_passwd}}{% else %}Oracle123{% endif%}"

However, as this is only set in the defaults/main.yml, the are overridden as soon as the new structure is used.
This check will also be removed at a later stage.