New module: oracle_redo - Manages redolog groups and their size
This module will ensure that you have a uniform redolog configuration.
You pass in the size of the redologs and how many groups you want, and then the module
ensures that state, regardless of how many threads you have (RAC or not).
However, it will only work on threads that are ENABLED.

If you are running a RAC configuration, the database needs to be in
ARCHIVELOG mode for the module to work. This is not required for single instance
environments.