oradb-manage-db: new handling of custom dbca-templates
The usage of dbca_templatename in oracle_database has been changed.
The template is copied to $ORACLE_HOME/assistant/dbca/templates for
each database as <db_name>_customexample.dbt. This allow the usage
of jinja2 in the template for each database.

Be aware:
The jinja2 variable 'item' has been changed to 'dbh' in the template
files.

oracle_databases:
   dbca_templatename: customexample.dbt