oraswdb-install: added support to multiple listener for database in filesystem
The stupid dbstart and dbstop has been replaced by a completly new
designed script manage_oracle_rdbms_procs.sh.
The script is placed in /usr/local/bin and must be executed as the
owner of the Oracle software.
Instances are started when the state in oratab is 'Y' or 'M' for only
'startup mount' which is important in data-guards to protect against
accidently using the active data-guard option.
All listeners are started when 'listener_name' is defined in oracle_databases
for the instance.

The script has some parameter:

manage_oracle_rdbms_procs.sh -a <start|stop> -m <abort|immediate>

The default is shutdown immediate This can be changed with the parameter
'-m abort'.