oradb-datapatch: extended support for datapatch
The role has been rewritten for Single-Instance configurations.
This role can be used to install PSU/RU for 12.1.0.2+ databases.

The following roles can be used to install a new PSU:
- oraswdb-manage-patches
- oradb-datapatch

The role starts listener and instances in SI configurations, because
oraswdb-manage-patches is stopping the listener and instances for the
ORACLE_HOME.

Don't forget the updates in ansible-oracle-modules for this commit.