chopt: new strategy for change chopt options
Oracle changed the logfile for chopt in 19c. This leads to problems with the old
impletation, due to recompiles of oracle binary for each execution of the task.
The new solution is not dependent on the logfile anymore.

Warning!
This change will recompile the binaries during the 1st execution, because
the statefile was not created with the old solution.