orahost: new approach for setting hugepages during setup
Some Kernels do not allow an increase of hugepages without
enough free memory. When hugepages could not be allocated
during initial setup the execution of DBCA could fail due
to missing enough hugpages.

When vm.nr_hugepages is changed the filecache is flushed
and anoher execution of systcl -p is done - hopefully to
allocate the needed pages.