oraswdb-install: Bugfix for mixed ORACLE_HOME installations
The change from include to include_task in 2018 broke the possibility to
install multiple ORACLE_HOMEs in parallel. This has been fixed.

I don't have the installation media for the following version.
Theses versions were not tested.
I'll remove 11.2.0.3 in the future. This version is out of support for
many years from Oracle and the addition of download passwords makes
it more worse to do tests with these versions.

- 11.2.0.3
- 12.1.0.1