ansible-oracle: BREAKING CHANGE! move to a collection (#99)
This is one of the biggest changes during the development of ansible-oracle since last refactoring.
The project moves to a collection, which leads to the following breaking changes:

- rename all roles with '-' in name to '_'
This is a mandatory change, when moving to a collection.
galaxy  will not import a collection with a '-' in rolename...
- Ansible 2.9+ needed
Due to the dependency to Collections, the minimum version of Ansible has been cahnged to 2.9.
- moved playbooks to playbooks folder
This is the 1st step to reorganize the old - mostly not working -
playbooks. The redesign of playbooks and example inventory is planned
for a later time.
- renamed variable for EE-Options in Binaries
All variables with `oracle_EE_options` have been changed to `oracle_ee_options`. An assertation has been added to check the inventory for old - unsupported - variables.


Other minor changes:
- update ansible-lint for git pre-commit to v6.3.0
- removed all check exceptions from .ansible-lint