oraemagent_install: removed role oraemagent_install
The role has beend removed, because it was really old and not working with current versions of Grid-Control anymore.
Some upcoming changes for ansible-doctor will create some rework in this role, due to centralizing of variables in a meta-role.
This will create some aditional work in this old - not working - role which is useless.
That is why this role is removed now.