cx_oracle: Added installation of cx_Oracle for Python3
Important:

The Release 3.8.0 of ansible-oracle changes the module oracle_db.
It requires an installation of Python3 and cx_Oracle for Python3.
The role cx_oracle installs cx_Oracle for Python2 and Python3
when 'ansible_python_interpreter' points to Python2.

Please execute the role cx_oracle for old systems to install cx_Oracle
for Python3.