Bugfix for missing Listener autostart and readonly Homes in systemd
The old dotprofile environment scripts were removed in 2022.
They were replaced by ocenv but still needed by manage_oracle_rdbms_procs.sh
for a mapping of listener to ORACLE_HOME.

The dotprofiles are back again with ocenv as environment switcher.

The start of database with readonly homes is supported now.