oraswgi: Added Support for gridSetup.sh -applyRU for 19c and 21c
The option to preinstall patches during gridSetup has been improved.
This change add the option '-applyRU' to the execution of gridSetup.sh
for 19c and 21c instead of the installation of patches with opatch
after the unarchive.
The installation of OneOff patches with `opatch apply` has been added
for 19c+, too.

The installation of patches with copy of patches from ansible controller or
download via http has been added, too.