This is a fix for #453 .
bugfix set custom environment for executables with oracle_script_env

`_JAVA_OPTIONS` is only set when `oracle_tmp_stage` != /tmp due to issues
with `GridSetup.sh -applyRU` for 21c.
The default for `oracle_tmp_stage` is /tmp when `ansible-fips` is disabled
or `{{ oracle_stage }}/tmp` when enabled.

The owner, group and priviledges for `oracle_tmp_stage` are set to same
values as `/tmp` on normal linux servers.

The `CV_ASSUME_DISTID` is set to `OL{{ ansible_distribution_major_version }}` for RHEL/OL
when not RHEL9/OL9 and set to `OL8` when RHEL9/OL9.

`SLES15` is default for SuSE at the moment. This could be changed in next PRs.

`cluvfy` is always executed without `CV_ASSUME_DISTID`, because the tool is
compatible for all supported plattforms when most current version is used.